flask_restful_api/requirements.txt
aniso8601==9.0.1
blinker==1.6.2
cffi==1.15.1
click==8.1.4
cryptography==41.0.2
Flask==2.3.2
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
mysqlclient==2.2.0
protobuf==3.20.3
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
Werkzeug==2.3.6
